% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsummit.R
\name{coordinate}
\alias{coordinate}
\title{Create coordinate objects}
\usage{
coordinate(longitude, latitude)
}
\arguments{
\item{longitude}{Longitude value.}

\item{latitude}{Latitude value.}
}
\value{
A coordinate object.
}
\description{
This function creates a coordinate object with longitude and latitude coordinates
}
