% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsummit.R
\name{create_region_counts_plot}
\alias{create_region_counts_plot}
\title{Plot number of mountains above 2500 ft by region
This function creates a bar plot showing the number of mountains above 2500 ft for each region.}
\usage{
create_region_counts_plot(data)
}
\arguments{
\item{data}{The merged data frame containing state and summit information.}
}
\value{
A ggplot object displaying the bar plot of regions.
}
\description{
Plot number of mountains above 2500 ft by region
This function creates a bar plot showing the number of mountains above 2500 ft for each region.
}
