% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsummit.R
\name{distance}
\alias{distance}
\title{Calculate distance between two coordinates}
\usage{
distance(coord1, coord2)
}
\arguments{
\item{coord1}{First coordinate.}

\item{coord2}{Second coordinate.}
}
\value{
Distance between the two coordinates in kilometers.
}
\description{
This function calculates the distance between two coordinates using Haversine formula.
}
