% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsummit.R
\name{print.itinerary}
\alias{print.itinerary}
\title{Print Itinerary}
\usage{
\method{print}{itinerary}(itinerary)

\method{print}{itinerary}(itinerary)
}
\arguments{
\item{x}{An itinerary object with the travel plan}

\item{destinations}{A list of coordinates representing custom destinations}
}
\value{
A list containing the generated travel itinerary
}
\description{
This function prints the itinerary.

This function generates an itinerary based on given destinations.
}
