% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsummit.R
\docType{data}
\name{state}
\alias{state}
\title{Load state data and merge with summit data}
\format{
An object of class \code{data.frame} with 15537 rows and 6 columns.
}
\usage{
state
}
\arguments{
\item{state_data}{The state data loaded from RStudio.}

\item{summit_data}{The summit data to be merged.}
}
\value{
Merged data frame with state and summit information.
}
\description{
This function loads state data from RStudio, renames columns, and merges it with USGS summit data.
}
\keyword{datasets}
