% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsummit.R
\docType{data}
\name{summits}
\alias{summits}
\title{Summit Data}
\format{
A data frame with variables:
\describe{
  \item{summit_name}{Name of the summit}
  \item{elevation}{Elevation of the summit}
  \item{latitude}{Latitude of the summit}
  \item{longitude}{Longitude of the summit}
}
}
\usage{
summits
}
\description{
A dataset containing information about summits.
}
\details{
Summit Data

This dataset contains information about summits.
}
\keyword{datasets}
